+++ tests/test_sql.py
@@ -19,7 +19,10 @@ from .factories.pluto_19v2 import Pluto19v2
 from .factories.real_property_master import RealPropertyMaster
 from .factories.real_property_legals import RealPropertyLegals
 
-import portfoliograph
+from portfoliograph.table import (
+    populate_portfolios_table,
+    export_portfolios_table_json
+)
 
 # This test suite defines two landlords:
 #
@@ -260,7 +263,7 @@ class TestSQL:
 
     def test_portfolio_graph_works(self):
         with self.db.connect() as conn:
-            portfoliograph.populate_portfolios_table(conn)
+            populate_portfolios_table(conn)
         r = self.query_one(
             "SELECT landlord_names FROM wow_portfolios WHERE '" + FUNKY_BBL + "' = any(bbls)"
         )
@@ -269,7 +272,7 @@ class TestSQL:
     def test_portfolio_graph_json_works(self):
         with self.db.connect() as conn:
             f = StringIO()
-            portfoliograph.export_graph_json(conn, f)
+            export_portfolios_table_json(conn, f)
 
             # Ideally we'd actually do a snapshot test here,
             # but I'm not confident that the ordering of
